Move special widget from colorimetry_blueprints.dart to its own specific file
