+++ lib/screens_pages/complete_profile_page.dart
@@ -11,8 +11,7 @@ import 'package:simple_chat/screens_pages/landing_page.dart';
 import 'package:simple_chat/utils/widgets_and_ui_elements/alert_dialog_list.dart';
 import 'package:simple_chat/utils/logger_config.dart';
 import 'package:simple_chat/session_related/app_storage_class.dart';
-
-import '../utils/colorimetry_blueprints.dart';
+import 'package:simple_chat/utils/widgets_and_ui_elements/labeled_text_field.dart';
 
 //imports
 /////////////////////////////////////////////////////////////////////////////
