+++ lib/screens_pages/settings_page.dart
@@ -5,11 +5,9 @@ import 'package:google_fonts/google_fonts.dart'; //Fonts
 import 'package:simple_chat/standalone_methods_functions/general_methods.dart';
 import 'package:simple_chat/configuration/config_invoke.dart';
 import 'package:simple_chat/utils/widgets_and_ui_elements/alert_dialog_list.dart';
-import 'package:simple_chat/utils/colorimetry_blueprints.dart';
-
+import 'package:simple_chat/utils/widgets_and_ui_elements/labeled_text_field.dart';
 import 'package:simple_chat/utils/logger_config.dart';
-
-import '../utils/easter_egg_player.dart';
+import 'package:simple_chat/utils/easter_egg_player.dart';
 
 class settings extends StatefulWidget {
   const settings({super.key});
