+++ lib/screens_pages/sign_up_page.dart
@@ -8,8 +8,7 @@ import 'package:simple_chat/configuration/config_invoke.dart';
 import 'package:simple_chat/utils/widgets_and_ui_elements/alert_dialog_list.dart';
 import 'package:simple_chat/utils/logger_config.dart';
 import 'package:simple_chat/screens_pages/log_in_page.dart';
-
-import '../utils/colorimetry_blueprints.dart';
+import 'package:simple_chat/utils/widgets_and_ui_elements/labeled_text_field.dart';
 
 //imports
 /////////////////////////////////////////////////////////////////////////////
